+++ api/v1/mdb/zz_generated.deepcopy.go
@@ -22,7 +22,7 @@ package mdb
 
 import (
 	"github.com/10gen/ops-manager-kubernetes/api/v1/status"
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )